pyproject.toml
[project]
name = "ai-python-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.2",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.38.32",
    "burr[start]>=0.40.2",
    "instructor>=1.8.3",
    "python-dotenv>=1.1.0",
    "strands-agents>=0.0.1",
    "strands-agents-builder>=0.0.1",
    "strands-agents-tools>=0.0.1",
]
